chore: pass useFipsEndpoint while computing getOutpostEndpoint
