+++ packages/middleware-sdk-s3-control/src/process-arnables-plugin/index.ts
@@ -1,7 +1,4 @@
 export { getProcessArnablesPlugin } from "./plugin";
 export { parseOutpostArnablesMiddleaware, parseOutpostArnablesMiddleawareOptions } from "./parse-outpost-arnables";
-export {
-  updateArnablesRequestMiddleware,
-  updateArnablesRequestMiddlewareOptions,
-  getOutpostEndpoint,
-} from "./update-arnables-request";
+export { updateArnablesRequestMiddleware, updateArnablesRequestMiddlewareOptions } from "./update-arnables-request";
+export { getOutpostEndpoint } from "./getOutpostEndpoint";
